+++ include/endstone/endstone.hpp
@@ -182,6 +182,7 @@ static_assert(_ITERATOR_DEBUG_LEVEL == 0,
 #include "inventory/equipment_slot.h"
 #include "inventory/exact_ingredient.h"
 #include "inventory/inventory.h"
+#include "inventory/item_complex_alias_ingredient.h"
 #include "inventory/item_factory.h"
 #include "inventory/item_stack.h"
 #include "inventory/item_tag_ingredient.h"