+++ src/endstone/core/inventory/item_metas.cpp
@@ -25,7 +25,8 @@ namespace endstone::core {
 EndstoneItemMetas::ItemMetaDetails EndstoneItemMetas::Empty = ItemMetaDetails::create<void>();
 EndstoneItemMetas::ItemMetaDetails EndstoneItemMetas::Item = ItemMetaDetails::create<EndstoneItemMeta>();
 EndstoneItemMetas::ItemMetaDetails EndstoneItemMetas::Map = ItemMetaDetails::create<EndstoneMapMeta>();
-EndstoneItemMetas::ItemMetaDetails EndstoneItemMetas::WritableBook = ItemMetaDetails::create<EndstoneWritableBookMeta>();
+EndstoneItemMetas::ItemMetaDetails EndstoneItemMetas::WritableBook =
+    ItemMetaDetails::create<EndstoneWritableBookMeta>();
 EndstoneItemMetas::ItemMetaDetails EndstoneItemMetas::Book = ItemMetaDetails::create<EndstoneBookMeta>();
 EndstoneItemMetas::ItemMetaDetails EndstoneItemMetas::Crossbow = ItemMetaDetails::create<EndstoneCrossbowMeta>();
 EndstoneItemMetas::ItemMetaDetails EndstoneItemMetas::Potion = ItemMetaDetails::create<EndstonePotionMeta>();