+++ src/endstone/core/inventory/recipe_ingredient.cpp
@@ -27,8 +27,10 @@ Nullable<endstone::RecipeIngredient> makeIngredient(std::shared_ptr<const ::Reci
     case ItemDescriptor::InternalType::Invalid:
         return nullptr;
     case ItemDescriptor::InternalType::ItemTag:
-    case ItemDescriptor::InternalType::ComplexAlias:
         return std::make_shared<EndstoneItemTagIngredient>(std::move(recipe), ingredient, descriptor.getFullName());
+    case ItemDescriptor::InternalType::ComplexAlias:
+        return std::make_shared<EndstoneItemComplexAliasIngredient>(std::move(recipe), ingredient,
+                                                                    descriptor.getFullName());
     case ItemDescriptor::InternalType::Default: {
         const auto *type = endstone::ItemType::get(descriptor.getFullName());
         if (descriptor.getItem() == nullptr || type == nullptr) {